refactor(catalog): centralize shared DTOs and interfaces for Car and Bike modules
Moved `CreateBrandDto`, `CreateModelDto`, `CreateTrimDto`, and `LiteItem` interface to a shared directory. Enhanced DTOs with Swagger decorators for improved API documentation. Updated respective service imports in Car and Bike modules.
