+++ ai-prompt/lgdlong/swagger-create-car-bike-catalog-dto.md
@@ -1,0 +1,122 @@
+# Prompt
+
+```aiignore
+import { IsInt, IsNotEmpty, IsOptional, IsString, MaxLength, Min } from 'class-validator';
+import { Type } from 'class-transformer';
+
+export class CreateBrandDto {
+  @IsString()
+  @IsNotEmpty()
+  @MaxLength(100)
+  name!: string;
+}
+
+export class CreateModelDto {
+  @IsString()
+  @IsNotEmpty()
+  @MaxLength(120)
+  name!: string;
+
+  // Dành cho endpoint POST /bike-catalog/models (không nested)
+  @IsOptional()
+  @Type(() => Number)
+  @IsInt()
+  @Min(1)
+  brandId?: number;
+}
+
+export class CreateTrimDto {
+  @IsString()
+  @IsNotEmpty()
+  @MaxLength(120)
+  name!: string;
+
+  // Dành cho endpoint POST /bike-catalog/trims (không nested)
+  @IsOptional()
+  @Type(() => Number)
+  @IsInt()
+  @Min(1)
+  modelId?: number;
+}
+
+do swagger for this dto
+```
+
+# Chat
+
+**Ok 👍 để Swagger hiển thị rõ schema cho các DTO này, bạn thêm các decorator `@ApiProperty()` của `@nestjs/swagger` vào từng field. Ví dụ như sau:
+
+```ts
+import { IsInt, IsNotEmpty, IsOptional, IsString, MaxLength, Min } from 'class-validator';
+import { Type } from 'class-transformer';
+import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
+
+export class CreateBrandDto {
+  @ApiProperty({
+    example: 'Toyota',
+    description: 'Tên hãng xe (tối đa 100 ký tự)',
+    maxLength: 100,
+  })
+  @IsString()
+  @IsNotEmpty()
+  @MaxLength(100)
+  name!: string;
+}
+
+export class CreateModelDto {
+  @ApiProperty({
+    example: 'Corolla',
+    description: 'Tên model (tối đa 120 ký tự)',
+    maxLength: 120,
+  })
+  @IsString()
+  @IsNotEmpty()
+  @MaxLength(120)
+  name!: string;
+
+  @ApiPropertyOptional({
+    example: 1,
+    description: 'ID của Brand (chỉ dùng cho endpoint POST /car-catalog/models)',
+    minimum: 1,
+  })
+  @IsOptional()
+  @Type(() => Number)
+  @IsInt()
+  @Min(1)
+  brandId?: number;
+}
+
+export class CreateTrimDto {
+  @ApiProperty({
+    example: '2.0 AT Premium',
+    description: 'Tên trim (tối đa 120 ký tự)',
+    maxLength: 120,
+  })
+  @IsString()
+  @IsNotEmpty()
+  @MaxLength(120)
+  name!: string;
+
+  @ApiPropertyOptional({
+    example: 10,
+    description: 'ID của Model (chỉ dùng cho endpoint POST /car-catalog/trims)',
+    minimum: 1,
+  })
+  @IsOptional()
+  @Type(() => Number)
+  @IsInt()
+  @Min(1)
+  modelId?: number;
+}
+```
+
+---
+
+✅ Giờ khi mở Swagger UI:
+
+* `CreateBrandDto` sẽ hiện field `name`.
+* `CreateModelDto` sẽ có `name` (bắt buộc) và `brandId` (optional).
+* `CreateTrimDto` sẽ có `name` (bắt buộc) và `modelId` (optional).
+
+Bạn có muốn mình viết thêm **response DTO** (`BrandDto`, `ModelDto`, `TrimDto`) với `id`, `createdAt`, `updatedAt`… để Swagger hiển thị kết quả trả về luôn không?
+**
