--- apps/api/src/modules/catalogs/cars/interfaces/lite-item.interface.ts
@@ -1,4 +1,0 @@
-export interface LiteItem {
-  id: number;
-  name: string;
-}
