+++ apps/api/src/modules/catalogs/shared/interfaces/lite-item.interface.ts
